[TC] Regenerated demo for tc_tc_capture_mode_sam_v71_xult
MH3-67828

+++ apps/tc/tc_capture_mode/firmware/src/config/sam_v71_xult/peripheral/tc/plib_tc0.c
@@ -97,13 +97,13 @@ uint32_t TC0_CH0_CaptureFrequencyGet( void )
 /* Read last captured value of Capture A */
 uint16_t TC0_CH0_CaptureAGet (void)
 {
-    return TC0_REGS->TC_CHANNEL[0].TC_RA;
+    return (uint16_t)TC0_REGS->TC_CHANNEL[0].TC_RA;
 }
 
 /* Read last captured value of Capture B */
 uint16_t TC0_CH0_CaptureBGet (void)
 {
-    return TC0_REGS->TC_CHANNEL[0].TC_RB;
+    return (uint16_t)TC0_REGS->TC_CHANNEL[0].TC_RB;
 }
 
 TC_CAPTURE_STATUS TC0_CH0_CaptureStatusGet(void)
